Feature/107 debenture event content

See merge request cash-flow-team/cash-flow!96

+++ client/lib/widgets/containers/game_event_selector.dart
@@ -124,7 +124,7 @@ class SelectorViewModel {
   });
 
   final double currentPrice;
-  final int passiveIncomePerMonth;
+  final double passiveIncomePerMonth;
   final int alreadyHave;
   final int maxCount;
   final bool changeableType;
